% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_15min.R
\name{az_15min}
\alias{az_15min}
\title{Retrieve 15-minute Weather Data from AZMet}
\source{
\url{https://azmet.arizona.edu/}
}
\usage{
az_15min(station_id = NULL, start_date_time = NULL, end_date_time = NULL)
}
\arguments{
\item{station_id}{Station ID can be supplied as numeric vector (e.g.
\code{station_id = c(8, 37)}) or as character vector with the prefix "az" and
two digits (e.g. \code{station_id = c("az08", "az37")}). If left blank, data for
all stations will be returned.}

\item{start_date_time}{A length-1 vector of class POSIXct or character in
YYYY-MM-DD HH:MM:SS format, in AZ time. If only a date (YYYY-MM-DD) is
supplied, data will be requested starting at 00:00:01 of that day.}

\item{end_date_time}{A length-1 vector of class POSIXct or character in
YYYY-MM-DD HH:MM:SS format, in AZ time. If only a date (YYYY-MM-DD) is
supplied, data will be requested through the \emph{end} of that day (23:59:59).
Defaults to the current date and time if left blank and \code{start_date_time}
is specified.}
}
\value{
A tibble. For units and other metadata, see
\url{https://azmet.arizona.edu/about}
}
\description{
Retrieves 15-minute data from the AZMet (Arizona Meteorological Network) API.
For a list of weather stations and their locations see \link{station_info}, or
visit https://azmet.arizona.edu/about.
}
\details{
If neither \code{start_date_time} nor \code{end_date_time} are supplied, the
most recent datetime of data will be returned. If only \code{start_date_time}
is supplied, then \code{end_date_time} defaults to the current time. Supplying
only \code{end_date_time} will result in an error.
}
\note{
If \code{station_id} is supplied as a vector, multiple successive calls to
the API will be made. You may find better performance getting data for all
the stations by leaving \code{station_id} blank and subsetting the resulting
dataframe. Only the most recent 48 hours of 15-minute data are stored in
the AZMet API.
}
\examples{
\dontrun{
# Most recent 15-minute data for all stations:
az_15min()

# Specify stations:
az_15min(station_id = c(1, 2))
az_15min(station_id = c("az01", "az02"))

# Specify dates:
az_15min(start_date_time = "2022-09-25 01:00:00")
az_15min(start_date_time = "2022-09-25 01:00:00", end_date_time = "2022-09-25 07:00:00")
}
}
\seealso{
\code{\link[=az_daily]{az_daily()}}, \code{\link[=az_heat]{az_heat()}}, \code{\link[=az_hourly]{az_hourly()}}, \code{\link[=az_lw15min]{az_lw15min()}}, \code{\link[=az_lwdaily]{az_lwdaily()}}
}
