% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_add_units.R
\name{az_add_units}
\alias{az_add_units}
\title{Assign units using the \code{units} package}
\usage{
az_add_units(x)
}
\arguments{
\item{x}{a tibble output by \code{\link[=az_hourly]{az_hourly()}}, \code{\link[=az_daily]{az_daily()}}, or \code{\link[=az_heat]{az_heat()}}}
}
\value{
a tibble with columns of class "units"
}
\description{
Assigns correct units to the output of \code{az_hourly()}, \code{az_daily()}, and
\code{az_heat()} using the \code{units} package.
}
\examples{
\donttest{
daily <- az_daily()

daily_units <-
  az_add_units(daily)

#unit conversions with `units::set_units()`
daily_units$sol_rad_total |> units::set_units("kW h m-2")

#units carry through calculations
climatic_balance <-
  daily_units$precip_total_mm - daily_units$eto_pen_mon
climatic_balance
}

}
