% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_heat.R
\name{az_heat}
\alias{az_heat}
\title{Retrieve Accumulated Heat Units and Evapotranspiration}
\source{
\url{https://azmet.arizona.edu/}
}
\usage{
az_heat(station_id = NULL, start_date = NULL, end_date = NULL)
}
\arguments{
\item{station_id}{station ID can be supplied as numeric vector (e.g.
\code{station_id = c(8, 37)}) or as character vector with the prefix "az" and 2
digits (e.g. \code{station_id = c("az08", "az37")}) If left blank data for all
stations will be returned}

\item{start_date}{A length-1 vector of class Date, POSIXct, or character in
YYYY-MM-DD format.  Will be rounded \strong{down} to the nearest day if more
precision is supplied.}

\item{end_date}{A length-1 vector of class Date, POSIXct, or character in
YYYY-MM-DD format.  Will be rounded \strong{down} to the nearest day if more
precision is supplied.  Defaults to the current date if left blank. If only
an \code{end_date} is supplied, then data will be cumulative from the start of
the year of \code{end_date}.}
}
\value{
A tibble. For units and other metadata, see
\url{https://azmet.arizona.edu/about}
}
\description{
Retrieves accumulated heat units and reference evapotranspiration units from
the Arizona Meteorological Network API. By default, returned values are
cumulative since January 1 of the current year. For a list of weather
stations and their locations see \link{station_info}, or visit
https://azmet.arizona.edu/about.
}
\details{
Unlike \code{\link[=az_daily]{az_daily()}}, only one row of data per station is returned,
regardless of \code{start_date} and \code{end_date}. However, the data returned is
cumulative over the time period specified by \code{start_date} and \code{end_date}.
}
\note{
If \code{station_id} is supplied as a vector, multiple successive calls to
the API will be made.  You may find better performance getting data for all
the stations by leaving \code{station_id} blank and subsetting the resulting
dataframe.
}
\examples{
\dontrun{
# Most recent data for all stations:
az_heat()

# Specify stations:
az_heat(station_id = c(1, 2))
az_heat(station_id = c("az01", "az02"))

# Specify dates:
## Cumulative since October 2022
az_heat(start_date = "2022-10-01")

## Cumulative from the first of the year through March
yr <- format(Sys.Date(), "\%Y")

az_heat(end_date = paste(yr, "03", "31", sep = "-"))
}

}
\seealso{
\code{\link[=az_15min]{az_15min()}}, \code{\link[=az_daily]{az_daily()}}, \code{\link[=az_hourly]{az_hourly()}}, \code{\link[=az_lw15min]{az_lw15min()}}, \code{\link[=az_lwdaily]{az_lwdaily()}}
}
