% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_hourly.R
\name{az_hourly}
\alias{az_hourly}
\title{Retrieve Hourly Weather Data}
\source{
\url{https://ag.arizona.edu/azmet/}
}
\usage{
az_hourly(station_id = NULL, start_date_time = NULL, end_date_time = NULL)
}
\arguments{
\item{station_id}{station ID can be supplied as numeric vector (e.g.
\code{station_id = c(8, 37)}) or as character vector with the prefix "az" and 2
digits (e.g. \code{station_id = c("az08", "az37")}) If left blank data for all
stations will be returned}

\item{start_date_time}{character; in YYYY-MM-DD HH or another format that can
be parsed by \code{\link[lubridate:ymd_hms]{lubridate::ymd_h()}}}

\item{end_date_time}{character; in YYYY-MM-DD in YYYY-MM-DD HH or another
format that can be parsed by \code{\link[lubridate:ymd_hms]{lubridate::ymd_h()}}.  Defaults to the current
time if left blank.}
}
\value{
a tibble
}
\description{
Retrieves hourly weather data from the Arizona Meteorological Network API.
For a list of weather stations and their locations see see \link{station_info}.
}
\details{
If neither \code{start_date_time} nor \code{end_date_time} are supplied, the
most recent day of data will be returned.  If only \code{start_date_time} is
supplied, then \code{end_date_time} defaults to the current time.  Supplying
only \code{end_date_time} will result in an error.
}
\note{
If \code{station_id} is supplied as a vector, multiple successive calls to
the API will be made.  You may find better performance getting data for all
the stations by leaving \code{station_id} blank and subsetting the resulting
dataframe.
}
\examples{
# Most recent data for all stations:
az_hourly()

# Specify stations:
az_hourly(station_id = c(1, 2))
az_hourly(station_id = c("az01", "az02"))

# Specify dates:
az_hourly(start_date_time = "2022-09-25 01")
az_hourly(start_date_time = "2022-09-25 01", end_date = "2022-09-25 20")

}
\seealso{
\code{\link[=az_daily]{az_daily()}}, \code{\link[=az_heat]{az_heat()}}
}
