% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_lwdaily.R
\name{az_lwdaily}
\alias{az_lwdaily}
\title{Retrieve Daily Leaf Wetness Data from AZMet}
\source{
\url{https://azmet.arizona.edu/}
}
\usage{
az_lwdaily(station_id = NULL, start_date = NULL, end_date = NULL)
}
\arguments{
\item{station_id}{station ID can be supplied as numeric vector (e.g.
\code{station_id = c(8, 37)}) or as character vector with the prefix "az" and
two digits (e.g. \code{station_id = c("az08", "az37")}). If left blank, data for
all stations will be returned.}

\item{start_date}{A length-1 vector of class Date, POSIXct, or character in
YYYY-MM-DD format. Will be rounded \strong{down} to the nearest day if more
precision is supplied. Defaults to the day before the current date (i.e.,
the most recent complete day) if left blank.}

\item{end_date}{A length-1 vector of class Date, POSIXct, or character in
YYYY-MM-DD format. Will be rounded \strong{down} to the nearest day if more
precision is supplied. Defaults to the day before the current date (i.e.,
the most recent complete day) if left blank.}
}
\value{
A tibble. For units and other metadata, see
\url{https://azmet.arizona.edu/about}
}
\description{
Retrieves daily leaf wetness data from the Arizona Meteorological Network
API. Currently, these data only are available from weather stations in the
Yuma area. For a list of stations and their locations see \link{station_info}, or
visit https://azmet.arizona.edu/about.
}
\details{
If neither \code{start_date} nor \code{end_date} are supplied, the most recent
day of data will be returned. If only \code{start_date} is supplied, then the
end date defaults to the day before the current date (i.e., the most recent
complete day). Supplying only \code{end_date} will result in an error.
}
\note{
If \code{station_id} is supplied as a vector, multiple successive calls to
the API will be made. You may find better performance getting data for all
the stations by leaving \code{station_id} blank and subsetting the resulting
dataframe. Requests for data from all stations for more than 6-12 months
may take considerable time.
}
\examples{
\dontrun{
# Most recent data for all stations:
az_lwdaily()

# Specify stations:
az_lwdaily(station_id = c(1, 2))
az_lwdaily(station_id = c("az01", "az02"))

# Specify dates:
az_lwdaily(start_date = "2022-09-25")
az_lwdaily(start_date = "2022-09-25", end_date = "2022-09-26")
}
}
\seealso{
\code{\link[=az_15min]{az_15min()}}, \code{\link[=az_daily]{az_daily()}}, \code{\link[=az_heat]{az_heat()}}, \code{\link[=az_hourly]{az_hourly()}}, \code{\link[=az_lw15min]{az_lw15min()}}
}
