% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_daily.R
\name{az_daily}
\alias{az_daily}
\title{Retrieve Daily Weather Data}
\usage{
az_daily(station_id = NULL, date_time_start = NULL, time_interval = NULL)
}
\arguments{
\item{station_id}{station ID. If left blank data for all stations will be returned}

\item{date_time_start}{When to start data return in YYYY-MM-DD HH:MM format.  If left blank the most recent day of data will be returned}

\item{time_interval}{time interval}
}
\value{
a data frame
}
\description{
Retrieve Daily Weather Data
}
\examples{
az_daily()
}
