% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_daily.R
\name{az_daily}
\alias{az_daily}
\title{Retrieve Daily Weather Data}
\usage{
az_daily(
  station_id = NULL,
  start_date = NULL,
  end_date = NULL,
  time_interval = NULL
)
}
\arguments{
\item{station_id}{station ID can be supplied as numeric vector (e.g.
\code{station_id = c(8, 37)}) or as character vector with the prefix "az" and 2
digits (e.g. \code{station_id = c("az08", "az37")}) If left blank data for all
stations will be returned}

\item{start_date}{character; in YYYY-MM-DD or another format that can be
parsed by \code{\link[lubridate:ymd]{lubridate::ymd()}}}

\item{end_date}{character; in YYYY-MM-DD in YYYY-MM-DD or another format that can be
parsed by \code{\link[lubridate:ymd]{lubridate::ymd()}}.  Defaults to the current date if left blank.}
}
\value{
a data frame
}
\description{
Retrieve Daily Weather Data
}
\details{
If neither \code{start_date} nor \code{end_date} are supplied, the most recent day of data will be returned.  If only \code{start_date} is supplied, then the end date defaults to the current date.  Supplying only \code{end_date} will result in an error.
}
\examples{
az_daily()
TODO: args should be start_date, end_date.  Do the parsing in the function I think.
}
