% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_info}
\alias{station_info}
\title{AZMet station names and locations}
\format{
A tibble with 30 rows and 5 columns:
\describe{
\item{\code{meta_station_name}}{Station name}
\item{\code{meta_station_id}}{Station ID}
\item{\code{latitude}, \code{logitude}}{Station location}
\item{\code{elev_m}}{Station elevation in meters}
}
}
\usage{
station_info
}
\description{
Station names, station IDs, and location (lat, lon, elevation) of active
AZMet weather stations
}
\keyword{datasets}
